Cont'd migration to data matching, added cluster support

+++ Source/PCGExtendedToolkit/Private/Data/PCGExPointIO.cpp
@@ -288,6 +288,12 @@ namespace PCGExData
 	{
 		PCGEX_SHARED_CONTEXT(ContextHandle)
 
+		if (InitOut == EIOInit::Forward && IsValid(Out) && Out == In)
+		{
+			// Already forwarding
+			return true;
+		}
+
 		if (IsValid(Out) && Out != In)
 		{
 			SharedContext.Get()->ManagedObjects->Destroy(Out);
